pyproject.toml
[project]
name = "tiro-reader"
version = "0.1.0"
description = "A local-first, open-source reading OS for the AI age"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "readability-lxml>=0.8.1",
    "markdownify>=0.14.1",
    "chromadb>=0.6.3",
    "sentence-transformers>=3.4.1",
    "anthropic>=0.43.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.2",
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.5",
    "python-multipart>=0.0.20",
]

[project.scripts]
tiro = "tiro.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["tiro"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
